+++ src/chrono_synchrono/controller/driver/SynMultiPathDriver.cpp
@@ -118,11 +118,11 @@ void ChMultiplePathSteeringController::Reset(const ChVehicle& vehicle) {
     }
 }
 
-int ChMultiplePathSteeringController::addPath(std::shared_ptr<ChBezierCurve> path, bool isClosedPath) {
+unsigned ChMultiplePathSteeringController::addPath(std::shared_ptr<ChBezierCurve> path, bool isClosedPath) {
     m_path.push_back(path);
     auto m_tracker_element = chrono_types::make_shared<ChBezierCurveTracker>(path, isClosedPath);
     m_tracker.push_back(m_tracker_element);
-    return m_tracker.size() - 1;
+    return (unsigned)(m_tracker.size() - 1);
 }
 
 }  // namespace synchrono
